backend/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.9
django-cors-headers==3.10.1
djangorestframework==3.13.1
psycopg2==2.9.2
pycodestyle==2.7.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
